Move NetworkManagerMock to a separate file
